Common.Object.SyntaxHighlight

Description

Documentation

asIf :: a -> a -> a

Evaluate to the first, with the second as a type hint.

coerceAsIf :: Coercible a b => (a -> b) -> a -> b

Coerce the data, as if it were converted using the given converter.

defaultTokenizerConfig :: TokenizerConfig

Default tokenizer config: with default syntax map, without trace output.

highlight :: Syntax -> Text -> [[SVG]]

Highlight with a specified language syntax. Use lookupSyntax to find a proper Syntax.

highlightIn :: Text -> Text -> [[SVG]]

Highlight code in some language.

highlightHs :: Text -> [[SVG]]

Highlight Haskell code.

highlightWith :: TokenizerConfig -> Style -> Syntax -> Text -> Either String [[SVG]]

Custom highlight with maximum flexibility.

escapeLaTeX :: Text -> Text

Escape LaTeX text. Stolen from Skylighting.Format.LaTeX.escapeLaTeX.