Common.Object.Transform

Description

Renderable Objects

SceneRender is used for procedural Object creation in place of Renderable.

class GroupRender a where

Allow rendering for groups of objects.

Methods

renderGroup :: Traversable t => t a -> t SVG

Render a group of objects to a group of SVGs. The group structure should be taken into consideration while rendering. e.g. Group of Text can be rendered using latexChunks.

oNewGroup :: forall a t s. (Traversable t, GroupRender a) => t a -> Scene s (t (Object s SVG))

Create a group of objects, translation of each lifted to Object level.

oNewGroupScaled :: forall a t s. (Traversable t, GroupRender a) => Double -> t a -> Scene s (t (Object s SVG))

Create a group of objects, scaled as a group, translation of each lifted to Object level.

oNewCentered :: SVG -> Scene s (Object s SVG)

New object, with its translation lifted to Object level.

Alignment

Alignment controls how the objects in the next key frame is aligned according to those objects from the current frame.

type GroupAlignment t t' s a b = t (Object s a) -> t' (Object s b) -> Scene s ()

Group alignment, given a base, align the second parameter according to the base.

applyAlignment :: (Traversable t, Traversable t') => [GroupAlignment t t' s a b] -> t (Object s a) -> t' (Object s b) -> Scene s ()

Apply some GroupAlignments.

Group Modifier

scaled :: (Traversable t, Traversable t') => Double -> GroupAlignment t t' s a b

Scaled by a factor.

allMargins :: Traversal (a, a, a, a) (b, b, b, b) a b

Traversal of all four margins.

withMargin :: (Traversable t, Traversable t') => Coord -> GroupAlignment t t' s a b

Set margin.

X Alignment

leftAligned :: (Traversable t, Traversable t') => GroupAlignment t t' s a b

Aligned according to the left boundary (minimum X).

rightAligned :: (Traversable t, Traversable t') => GroupAlignment t t' s a b

Aligned according to the right boundary (maximum X).

xCentered :: (Traversable t, Traversable t') => GroupAlignment t t' s a b

Aligned according to the horizontal center line (average of min/max X).

placedBelow :: (Traversable t, Traversable t') => GroupAlignment t t' s a b

Placed below the alignment base.

Y Alignment

bottomAligned :: (Traversable t, Traversable t') => GroupAlignment t t' s a b

Aligned according to the bottom boundary (minimum Y).

topAligned :: (Traversable t, Traversable t') => GroupAlignment t t' s a b

Aligned according to the top boundary (minimum Y).

yCentered :: (Traversable t, Traversable t') => GroupAlignment t t' s a b

Aligned according to the vertical center line (average of min/max Y).

Key Frame Animations

data Trans a b

Description for key frame animations.

Note:

IsString instance provided here for conveniently creating new objects.

The following objects are collected as the base for alignment:

Objects that vanished, or

Source objects for transformations

The following objects are collected as the target for alignment:

Objects newly created, or

Target objects for transformations

Constructors

New b
Object newly created in this frame.

Vanish a
Previous object vanishing in this frame.

[a] ::=> b
Object transformaton in this frame.

Bundled Patterns

pattern (:=>) :: a -> b -> Trans a b
Short cut for single object transformaton in this frame.

transOrigin :: Traversal (Trans a b) (Trans a' b) a a'

Traverse the origins of the Trans.

transformObject' :: GroupRender t => [GroupAlignment (Compose [] []) (Compose [] (Trans (Object s a))) s a SVG] -> [Trans (Object s a) t] -> Scene s [Object s SVG]

Key frame animation, with automatic creation for new objects.

Warning: if alignment base (see Trans for details) is empty, then all the alignment functions provided in this module would crash your program. In this case, custom (ad-hoc) alignment functions might be useful. Or, to specify a custom alignment base bs, use const (alignment bs) instead of alignment.

Note: if the target object utilise OverloadedStrings via IsString, the type becomes ambiguous. Use TypeApplications for this situation:

transformObject' @ObjType alignments transformations

transformObject :: [Trans (Object s a) [Object s b]] -> Scene s ()

Key frame animation.

Warning: if alignment base (see Trans for details) is empty, then all the alignment functions provided in this module would crash your program. In this case, custom (ad-hoc) alignment functions might be useful. Or, to specify a custom alignment base bs, use const (alignment bs) instead of alignment.

dupFromObject :: Renderable a => [Trans (Object s a) [Object s b]] -> Scene s ()

Key frame animation, but without hiding the original.

Object duplication

dupObject :: Renderable a => Object s a -> Scene s (Object s a)

Duplicate the object, make it independent from the original.

Primitives

readGroupTrans :: Traversable t => t (Object s a) -> Scene s (V2 Double)

Read the overall translation of an Object group.