Common.SVG

Description

Documentation

adjustStrokeWidth :: (Double -> Double) -> SVG -> SVG

Adjust the strokeWidth for an SVG.

counterclockwiseLR :: [PathCommand] -> [PathCommand]

Predefined style for paths: every path is counterclockwise, and left paths appear first.

clockwiseLR :: [PathCommand] -> [PathCommand]

Predefined style for paths: every path is clockwise, and left paths appear first.

groupPath :: [PathCommand] -> [[PathCommand]]

Group a series of PathCommands by subpaths.

sortPathBy :: ([PathCommand] -> [PathCommand] -> Ordering) -> [PathCommand] -> [PathCommand]

Sort subpaths by some ordering.

sortPathOn :: Ord a => ([PathCommand] -> a) -> [PathCommand] -> [PathCommand]

Sort subpaths by comparing the results of a key function applied to each element.

sortPathOn f is equivalent to sortPathBy (comparing f), but has the performance advantage of only evaluating f once for each element in the input list. This is called the decorate-sort-undecorate paradigm, or Schwartzian transform.

See also sortOn and sortBy.

For predefined key functions, see pathBB, pathCenter, pathRadius.

pathBB :: [PathCommand] -> (Coord, Coord, Coord, Coord)

Bounding box of a (sub-) path.

pathCenter :: [PathCommand] -> RPoint

Center of a path.

pathRadius :: [PathCommand] -> Coord

Radius of a (sub-) path, i.e. maximum distance from the center of the bounding box.

getPoints :: [LineCommand] -> [RPoint]

Get end points and control points for a path (as [LineCommand]).

flipLine :: [LineCommand] -> [LineCommand]

Flip the directon of an SVG path (as [LineCommand]). Prefer flipPath over roundtripping through toLineCommands and lineToPath.

isClockwise :: [RPoint] -> Bool

Check whether a (closed) path is clockwise or counterclockwise.

makeClockwise :: [PathCommand] -> [PathCommand]

Make a path clockwise by flipping if it was not.

makeCounterclockwise :: [PathCommand] -> [PathCommand]

Make a path clockwise by flipping if it was not.

flipPath :: [PathCommand] -> [PathCommand]

Flip the directon of an SVG path.